Add mobile variants table
