+++ packages/app/src/domain/variants/variants-table-tile/components/index.tsx
@@ -1,0 +1,7 @@
+export { DesktopVariantsTable } from './desktop-variants-table';
+export { MobileVariantsTable } from './mobile-variants-table';
+export { PercentageBarWithNumber } from './percentage-bar-with-number';
+export { Samples } from './samples';
+export { Cell, HeaderCell, StyledTable } from './shared-table-components';
+export { VariantCell } from './variant-cell';
+export { VariantDiff } from './variant-diff';
