+++ packages/app/src/domain/variants/variants-table-tile/components/percentage-bar-with-number.tsx
@@ -1,0 +1,27 @@
+import { Box } from '~/components/base';
+import { PercentageBar } from '~/components/percentage-bar';
+import { InlineText } from '~/components/typography';
+import { useIntl } from '~/intl';
+
+export function PercentageBarWithNumber({
+  percentage,
+  color,
+}: {
+  percentage: number;
+  color: string;
+}) {
+  const { formatPercentage } = useIntl();
+  return (
+    <Box
+      color={color}
+      display="flex"
+      alignItems="center"
+      pr={{ _: 2, sm: 2, lg: 4, xl: 5 }}
+    >
+      <InlineText fontWeight="bold" color="black" pr={2}>
+        {`${formatPercentage(percentage)}%`}
+      </InlineText>
+      <PercentageBar percentage={percentage} height="8px" />
+    </Box>
+  );
+}
