+++ packages/app/src/domain/variants/variants-table-tile/components/samples.tsx
@@ -1,0 +1,19 @@
+import { InlineText } from '~/components/typography';
+import { useIntl } from '~/intl';
+
+export function Samples({
+  occurrence,
+  sampleSize,
+}: {
+  occurrence: number;
+  sampleSize: number;
+}) {
+  const { formatNumber } = useIntl();
+
+  return (
+    <>
+      <InlineText fontWeight="bold">{formatNumber(occurrence)}</InlineText>
+      <InlineText>/{formatNumber(sampleSize)}</InlineText>
+    </>
+  );
+}
