+++ packages/app/src/domain/variants/variants-table-tile/components/variant-cell.tsx
@@ -1,0 +1,27 @@
+import { InlineTooltip } from '~/components/inline-tooltip';
+import { InlineText } from '~/components/typography';
+import { SiteText } from '~/locale';
+import { Cell } from '.';
+import { Variant } from '../logic/use-variants-table-data';
+
+type VariantCellProps = {
+  variant: Variant;
+  text: SiteText['covid_varianten'];
+  compact?: boolean;
+};
+
+export function VariantCell(props: VariantCellProps) {
+  const { variant, text, compact } = props;
+
+  return (
+    <Cell compact={compact}>
+      {variant === 'other' ? (
+        <InlineTooltip content={text.varianten_tabel.anderen_tooltip}>
+          <InlineText fontWeight="bold">{text.varianten[variant]}</InlineText>
+        </InlineTooltip>
+      ) : (
+        <InlineText fontWeight="bold">{text.varianten[variant]}</InlineText>
+      )}
+    </Cell>
+  );
+}
